esclude tests from build

+++ src/dcct/fftw_actuator.cc
@@ -7,7 +7,7 @@ Eigen::MatrixXd dcct::FFTWActuator::dct(Eigen::MatrixXd& X) {
   double_t sqrt_of_one_half_over_N = std::sqrt(0.5 / (double_t) M);
   double_t one_over_sqrt_of_two = std::sqrt(0.5);
 
-  int dims[1] = {M}; 
+  int dims[1] = {(int)M}; 
   fftw_r2r_kind kinds[1] = {FFTW_REDFT10};
   fftw_plan plan = fftw_plan_many_r2r(1, dims, N, const_cast<double*>(X.data()), dims, N, 1, Y.data(), dims, N, 1, kinds, FFTW_PRESERVE_INPUT);
   fftw_execute_r2r(plan, const_cast<double*>(X.data()), Y.data());
@@ -33,7 +33,7 @@ Eigen::MatrixXd dcct::FFTWActuator::idct(Eigen::MatrixXd& Y) {
   Y /= sqrt_of_one_half_over_N;
   Y.col(0) /= one_over_sqrt_of_two;
 
-  int dims[1] = {M}; 
+  int dims[1] = {(int)M}; 
   fftw_r2r_kind kinds[1] = {FFTW_REDFT01};
   fftw_plan plan = fftw_plan_many_r2r(1, dims, N, Y.data(), dims, N, 1, X.data(), dims, N, 1, kinds, FFTW_DESTROY_INPUT);
   fftw_execute_r2r(plan, Y.data(), X.data());
